wip todo implementation

+++ src/hooks/useTodo.ts
@@ -1,17 +1,20 @@
-import { useState } from "react";
-import { CreatTodoPayloadAttr, createTodo, deleteTodo } from "../../store/features/todoSlice";
-import { useAppDispatch } from "../../store/store";
+import { useEffect, useState } from "react";
+import { CreatTodoPayloadAttr, createTodo, deleteTodo, getTodo } from "../../store/features/todoSlice";
+import { useAppDispatch, useAppSelector } from "../../store/store";
 
 export const useTodo = () => {
 	const [form, setForm] = useState<CreatTodoPayloadAttr>();
 	const [error, setError] = useState<CreatTodoPayloadAttr>();
+	const { isCreatingTodo, isDeletingTodo, isLoading, todos, todo } = useAppSelector((store) => store.todo);
 	const dispatch = useAppDispatch();
 	const handleChange = (name: string, val: string) => {
-		if (!val) {
-			return setError((prev: CreatTodoPayloadAttr) => ({
+		if (val == "") {
+			setError((prev: CreatTodoPayloadAttr) => ({
 				...prev,
 				[name]: `${name} is required`,
 			}));
+			setForm(null);
+			return;
 		}
 		setError(null);
 		setForm((prev: CreatTodoPayloadAttr) => ({
@@ -20,14 +23,20 @@ export const useTodo = () => {
 		}));
 	};
 
+	useEffect(() => {
+		dispatch(getTodo());
+	}, []);
+
 	const handleCreateTodo = () => {
 		const object = { ...form, status: "active", archived: false, priority: "high" };
-		console.log(object);
-		dispatch(createTodo(object));
+		dispatch(createTodo(object)).then(() => {});
 	};
 
 	const deletedTodo = (todoId: string) => {
-		dispatch(deleteTodo(todoId));
+		console.log(todoId);
+		if (window.confirm("Are you sure you want to delete this todo")) {
+			dispatch(deleteTodo(todoId));
+		}
 	};
 	return {
 		form,
@@ -37,5 +46,10 @@ export const useTodo = () => {
 		handleChange,
 		handleCreateTodo,
 		deletedTodo,
+		isCreatingTodo,
+		isDeletingTodo,
+		isLoading,
+		todos,
+		todo,
 	};
 };
